Add YCKP feature: models, controllers, and views
Introduced new models (C242_YCKP_Files, C242_YCKP_New, C242_YCKP_Response, View_242_YCKPXL) and corresponding API controllers for handling 'Yêu Cầu Khắc Phục' (YCKP) requests, file uploads, and responses. Updated NN_Database context and EDMX to include new entities. Added new views for YCKP processing in the ProcessingDept area. Also included minor refactoring and new endpoints in existing controllers to support related data queries.

+++ NhatNam_web_new/NNworking/NNworking/Models/sp_Get_StaffCapability_Result.cs
@@ -20,10 +20,10 @@ namespace NNworking.Models
         public Nullable<double> ttgpbjpbn { get; set; }
         public Nullable<double> ttgpbjttc { get; set; }
         public Nullable<decimal> ttgk { get; set; }
-        public Nullable<double> C_B1 { get; set; }
+        public Nullable<decimal> C_B1 { get; set; }
         public Nullable<decimal> C_C1 { get; set; }
         public Nullable<double> C_D1 { get; set; }
         public Nullable<double> C_E1 { get; set; }
-        public Nullable<double> C_G1 { get; set; }
+        public Nullable<decimal> C_G1 { get; set; }
     }
 }
